skip create .gitignore if exist

+++ src/utils/files/gitignore.js
@@ -6,6 +6,9 @@ import {USERJSON} from '../../params';
 export const GITIGNORE = ".gitignore"
 
 export const createGitIgnore = (repository) => {
+    if(fs.existsSync(path.resolve(repository, GITIGNORE))){
+        return;
+    }
     const content = `${USERJSON}`;
     fs.writeFileSync(path.resolve(repository, GITIGNORE), content);
 }
